Add files via upload
a Python function called add_sub that takes two input arguments x and y. The function first adds x and y and stores the result in a variable called a. It then subtracts y from x and stores the result in a variable called b. Finally, it returns both a and b as a tuple.


# The function can be called with two arguments, as shown in the line result1,result2 = add_sub(6,7). This line assigns the values returned by the add_sub function to the variables result1 and result2, respectively. The print statement then outputs the values of result1 and result2 to the console.